[this commit is not completed , but for commit log] feat : Socket Class Creates
Socket Class is a WRapper class.

Socket Class contains SOCKADDR_IN data.

feat : Session Class and Session Manager Class Creates
A session class is a class that contains Client Session information.
Inherit the Session Class to implement multiple sessions.

The Session Manager Class is assigned through pools and pools for the Session.
It has a list of sessions and manages sessions on the IOCP server.
It is a class used for.

It still needs to implement Acceptex, DisconnectEx, and so on.